move blog code to separate file
